+++ medisupply_movil/lib/data/order_item.dart
@@ -2,6 +2,6 @@ class OrderItem {
   final int productoId;
   final String nombre;
   final int cantidad;
-  final int precio;
+  final double precio;
   OrderItem({required this.productoId, required this.nombre, required this.cantidad, required this.precio});
 }
