+++ medisupply_movil/lib/data/product.dart
@@ -1,7 +1,7 @@
 class Product {
   final int id;
   final String nombre;
-  final int precio;
+  final double precio;
   final int stock;
   final String categoria;
   Product({required this.id, required this.nombre, required this.precio, required this.stock, required this.categoria});
